Refactored some core components
