+++ HABApp/__version__.py
@@ -1,1 +1,1 @@
-__VERSION__ = '0.5.0'
+__VERSION__ = '0.5.1'
