+++ HABApp/core/__init__.py
@@ -1,7 +1,8 @@
 from .wrappedfunction import WrappedFunction
 
-from .events import EventBus as __EventBus
-from .events import EventListener, ValueUpdateEvent, ValueChangeEvent, ValueNoChangeEvent, ValueNoUpdateEvent, AllEvents
+from .event_bus import EventBus as __EventBus
+from .event_bus_listener import EventBusListener
+from .events import ValueUpdateEvent, ValueChangeEvent, ValueNoChangeEvent, ValueNoUpdateEvent, AllEvents
 from .items import Items as __Items
 from .items import Item
